[RENAME/#20] 유저 DAO 패키지명 변경
