+++ src/main/java/com/togedy/togedy_server_v2/domain/user/dao/UserRepository.java
@@ -1,4 +1,4 @@
-package com.togedy.togedy_server_v2.domain.user.dto;
+package com.togedy.togedy_server_v2.domain.user.dao;
 
 import com.togedy.togedy_server_v2.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
